pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools-scm>=8"]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]
# can be empty if no extra settings are needed, presence enables setuptools-scm

[project.urls]
Homepage = "https://github.com/CQCL/QUARK-plugin-quantinuum"

[project]
name = "QUARK-plugin-quantinuum"
dynamic = ["version"]
authors = [
    {name = "Etienne Granet", email = "etienne.granet@quantinuum.com"},
    {name = "Travis Thompson", email = "travis.thompson@quantinuum.com"},
]
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
license = "Apache-2.0"
dependencies = [
    "qiskit>=1.3.0",
    "numpy>=1.26.4",
    "matplotlib>=3.9.3",
    "qiskit>=1.3.0",
    "qiskit_aer>=0.15.1",
    "scipy>=1.12.0",
    "quark-framework>=0.3.7",
    "qnexus>=0.37.0",
    "pytket>=2.10.2",
    "pytket-qiskit>=0.72.0",
]
[dependency-groups]
dev = [
    "pytest (>=8.0.0,<9.0.0)",
    "ruff (>=0.12.0,<0.13.0)",
    "mypy >=1.9.0,<2",
    "scipy-stubs>=1.16.2.0",
]
